requirements.txt
python-dotenv
python-telegram-bot
web3
requests